:art: refactor : 패키지 변경 가능이라는 사항이 안적혀있는 파일들의 위치 원상 복구
