+++ src/test/java/bridge/BridgeNumberGeneratorTest.java
@@ -1,5 +1,6 @@
-package bridge.domain.strategy;
+package bridge;
 
+import bridge.BridgeNumberGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
